# docs/swagger.yaml
definitions:
  domain.Account:
    properties:
      account_id:
        type: integer
      document_number:
        type: string
    type: object
  domain.Transaction:
    properties:
      account_id:
        type: integer
      amount:
        type: number
      event_date:
        type: string
      operation_type_id:
        type: integer
      transaction_id:
        type: integer
    type: object
  handler.CreateAccountRequest:
    properties:
      document_number:
        type: string
    required:
    - document_number
    type: object
  handler.CreateTransactionRequest:
    properties:
      account_id:
        type: integer
      amount:
        type: number
      operation_type_id:
        type: integer
    required:
    - account_id
    - amount
    - operation_type_id
    type: object
info:
  contact: {}
paths:
  /accounts:
    post:
      consumes:
      - application/json
      description: Cria uma conta baseada no número do documento
      parameters:
      - description: Dados da conta
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/handler.CreateAccountRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Account'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Cria uma nova conta
      tags:
      - accounts
  /accounts/{accountId}:
    get:
      consumes:
      - application/json
      description: Retorna os detalhes de uma conta pelo ID
      parameters:
      - description: ID da conta
        in: path
        name: accountId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Account'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Consulta uma conta
      tags:
      - accounts
  /transactions:
    post:
      consumes:
      - application/json
      description: Registra uma transação e inverte o sinal baseado no tipo
      parameters:
      - description: Dados da transação
        in: body
        name: transaction
        required: true
        schema:
          $ref: '#/definitions/handler.CreateTransactionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Transaction'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Cria uma nova transação
      tags:
      - transactions
swagger: "2.0"